# Remove intermediary reader and directly feed iterator for unmarshal